requirements.txt
hmmlearn>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
scikit-learn>=1.3.0
ta>=0.11.0
scipy>=1.11.0
tqdm>=4.65.0
